refactor: move icons to /lib
